# [Feature](inverted index) support range predicate for inverted index