--- a/be/src/common/config.cpp
+++ b/be/src/common/config.cpp
@@ -1011,6 +1011,7 @@ DEFINE_Int32(inverted_index_read_buffer_size, "4096");
 DEFINE_Int32(max_depth_in_bkd_tree, "32");
 // index compaction
 DEFINE_Bool(inverted_index_compaction_enable, "false");
+DEFINE_mInt32(inverted_index_max_terms, "1024");
 // use num_broadcast_buffer blocks as buffer to do broadcast
 DEFINE_Int32(num_broadcast_buffer, "32");
 // semi-structure configs