… be/src/olap/rowset/segment_v2/inverted_index/query/inverted_index_query.cpp
… be/src/olap/rowset/segment_v2/inverted_index/query/range_query.cpp
… be/src/olap/rowset/segment_v2/inverted_index_reader.cpp
… be/src/olap/rowset/segment_v2/inverted_index_reader.h
… be/src/olap/rowset/segment_v2/segment_iterator.cpp
… be/src/pipeline/exec/olap_scan_operator.h
… be/src/vec/exec/scan/new_olap_scan_node.cpp
… be/src/vec/exec/scan/new_olap_scan_node.h
… be/src/vec/exec/scan/new_olap_scanner.cpp
… be/src/vec/exec/scan/vscan_node.h