## Proposed changes

Issue Number: close #xxx

Test with httplogs dataset, schema as following, we use inverted index for size column:
```
CREATE TABLE `httplogs` (
  `@timestamp` int(11) NULL,
  `clientip` varchar(20) NULL,
  `request` text NULL,
  `status` int(11) NULL,
  `size` int(11) NULL,
  INDEX size_idx (`size`) USING INVERTED COMMENT ''
) ENGINE=OLAP
DUPLICATE KEY(`@timestamp`)
COMMENT 'OLAP'
DISTRIBUTED BY HASH(`@timestamp`) BUCKETS 24
PROPERTIES (
"replication_allocation" = "tag.location.default: 1",
"is_being_synced" = "false",
"storage_format" = "V2",
"compression" = "ZSTD",
"light_schema_change" = "true",
"disable_auto_compaction" = "false",
"enable_single_replica_compaction" = "false"
);
```

Total lines for httplogs:
```
mysql> select COUNT() from httplogs;
+------------+
| count(*)   |
+------------+
| 1474712832 |
+------------+
1 row in set (0.18 sec)
```

before optimization:
```
mysql> select COUNT() from httplogs where size > 20000 and size <= 20001;
+----------+
| count(*) |
+----------+
|      150 |
+----------+
1 row in set (1.33 sec)
```

after range query optimization:
```
mysql> select COUNT() from httplogs where size > 20000 and size <= 20001;
+----------+
| count(*) |
+----------+
|      150 |
+----------+
1 row in set (0.05 sec)
```

Without inverted index:

```
mysql> set enable_inverted_index_query=false;                                                                                                                             Query OK, 0 rows affected (0.01 sec)

mysql> select COUNT() from httplogs where size > 20000 and size <= 20001;
+----------+
| count(*) |
+----------+
|      150 |
+----------+
1 row in set (0.62 sec)
```
## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...