@airborne12 2023-09-18
run buildall