@airborne12 2023-10-23
run clickbench